Add test project and extend Maybe functionality
A new NUnit test project, DiscriminatedUnionsTests, has been added to the solution. This contains tests for two newly introduced methods in the Calculator class of the CI project: Divide_ValidInputs_ReturnsCorrectResult and Divide_DivideByZero_ReturnsNothing. The Maybe.cs class of the CI project got extended with a Bind method which forms an important part of the monadic design for the class. New tests to check this method were also written. This will ensure that the Calculator and Maybe classes function as expected and further development can be done with a safety net of tests.