Refactor: Made `Calculator.Divide` a static method
Removed the setup of a Calculator instance in tests and in Main of Program.cs, instead calling `Calculator.Divide` as a static method. This simplifies the use of the `Divide` method, as an instance of the `Calculator` class doesn't need to be created for every division operation. This change was also propagated to the tests, removing the need to have a `SetUp` method for instantiating a Calculator object before each test.

Also, removed unnecessary 'this' keyword in Maybe.cs.