Add overloads to Bind function for different types
Updated the Bind function in MaybeExtensions to support reference types, non-nullable value types, and nullable value types with the BindRef, BindVal, and BindNullable methods respectively. This is to ensure that our generic Maybe Monad behaves correctly for all types of data. This change will lead to less unexpected behavior, making the Monad more robust and reliable. Tests have been added to ensure that the newly implemented behavior works as expected.